+++ federation/tests/entities/diaspora/test_mappers.py
@@ -18,7 +18,8 @@ from federation.tests.fixtures.payloads import (
     DIASPORA_POST_WITH_PHOTOS, DIASPORA_CONTACT,
     DIASPORA_PROFILE_EMPTY_TAGS, DIASPORA_RESHARE,
     DIASPORA_RESHARE_WITH_EXTRA_PROPERTIES, DIASPORA_POST_SIMPLE_WITH_MENTION,
-    DIASPORA_PROFILE_FIRST_NAME_ONLY, DIASPORA_POST_COMMENT_NESTED)
+    DIASPORA_PROFILE_FIRST_NAME_ONLY, DIASPORA_POST_COMMENT_NESTED, DIASPORA_POST_ACTIVITYPUB_ID,
+    DIASPORA_POST_COMMENT_ACTIVITYPUB_ID, DIASPORA_PROFILE_ACTIVITYPUB_ID)
 from federation.types import UserType, ReceiverVariant
 
 
@@ -30,6 +31,16 @@ class TestDiasporaEntityMappersReceive:
         assert len(entities) == 1
         post = entities[0]
         assert post._mentions == {'jaywink@jasonrobinson.me'}
+
+    def test_message_to_objects_post__with_activitypub_id(self):
+        entities = message_to_objects(DIASPORA_POST_ACTIVITYPUB_ID, "alice@alice.diaspora.example.org")
+        assert len(entities) == 1
+        post = entities[0]
+        assert isinstance(post, DiasporaPost)
+        assert isinstance(post, Post)
+        assert post.guid == "((guidguidguidguidguidguidguid))"
+        assert post.handle == "alice@alice.diaspora.example.org"
+        assert post.id == "https://alice.diaspora.example.org/posts/1"
 
     def test_message_to_objects_simple_post(self):
         entities = message_to_objects(DIASPORA_POST_SIMPLE, "alice@alice.diaspora.example.org")
@@ -78,6 +89,21 @@ class TestDiasporaEntityMappersReceive:
         assert comment._xml_tags == [
             "guid", "parent_guid", "text", "author",
         ]
+        mock_validate.assert_called_once_with()
+
+    @patch("federation.entities.diaspora.mappers.DiasporaComment._validate_signatures")
+    def test_message_to_objects_comment__activitypub_id(self, mock_validate):
+        entities = message_to_objects(DIASPORA_POST_COMMENT_ACTIVITYPUB_ID, "alice@alice.diaspora.example.org",
+                                      sender_key_fetcher=Mock())
+        assert len(entities) == 1
+        comment = entities[0]
+        assert isinstance(comment, DiasporaComment)
+        assert isinstance(comment, Comment)
+        assert comment.target_guid == "((parent_guidparent_guidparent_guidparent_guid))"
+        assert comment.root_target_guid == ""
+        assert comment.guid == "((guidguidguidguidguidguid))"
+        assert comment.handle == "alice@alice.diaspora.example.org"
+        assert comment.id == "https://alice.diaspora.example.org/comments/1"
         mock_validate.assert_called_once_with()
 
     @patch("federation.entities.diaspora.mappers.DiasporaComment._validate_signatures")
@@ -140,6 +166,16 @@ class TestDiasporaEntityMappersReceive:
         assert profile.public == True
         assert profile.nsfw == False
         assert profile.tag_list == ["socialfederation", "federation"]
+
+    @patch("federation.entities.diaspora.mappers.retrieve_and_parse_profile", return_value=Mock(
+        id="bob@example.com",
+    ))
+    def test_message_to_objects_profile__activitypub_id(self, mock_parse):
+        entities = message_to_objects(DIASPORA_PROFILE_ACTIVITYPUB_ID, "bob@example.com")
+        assert len(entities) == 1
+        profile = entities[0]
+        assert profile.handle == "bob@example.com"
+        assert profile.id == "https://example.com/bob"
 
     @patch("federation.entities.diaspora.mappers.retrieve_and_parse_profile", return_value=Mock(
         id="bob@example.com",